+++ src/endpoints/fetcher/fetcher_test.go
@@ -4,7 +4,7 @@ import "testing"
 
 func Test_GetListOfEndpoints(t *testing.T) {
 	var EXPECTED_ENDPOINTS = 354
-	var endpoints = GetListOfEndpoints("../../resources/EndpointSources.json")
+	var endpoints = GetListOfEndpoints("../../../resources/EndpointSources.json")
 	var endpointsCount = len(endpoints.Entries)
 	if endpointsCount != EXPECTED_ENDPOINTS {
 		t.Errorf("Number of endpoints read from resource file incorrect, got: %d, want: %d.", endpointsCount, EXPECTED_ENDPOINTS)
